+++ CanMessageSender.h
@@ -50,7 +50,7 @@ class CanMessageSender : public AbstractCanMessageSender {
           ESP_ERROR_CHECK(can_start());
       }
 
-      virtual void SendMessage(unsigned long canId, uint8_t ext, uint8_t sizeOfByteArray, unsigned char *byteArray)
+      virtual void SendMessage(uint16_t canId, uint8_t ext, uint8_t sizeOfByteArray, unsigned char *byteArray)
       {
           can_message_t tx_msg;
           tx_msg.data_length_code = sizeOfByteArray;
@@ -72,7 +72,7 @@ class CanMessageSender : public AbstractCanMessageSender {
 
       }
 
-      virtual void ReadMessage(uint32_t* canId, uint8_t* len, uint8_t* buf)
+      virtual void ReadMessage(uint16_t* canId, uint8_t* len, uint8_t* buf)
       {
           can_message_t rx_msg;
           ESP_ERROR_CHECK(can_receive(&rx_msg, portMAX_DELAY));
