+++ CanRadioTunerStructs.h
@@ -5,7 +5,7 @@
     #define _CanRadioTunerStructs_h
 
 // CANID: 225
-const unsigned long CAN_ID_RADIO_TUNER = 0x225;
+const uint16_t CAN_ID_RADIO_TUNER = 0x225;
 
 const uint8_t CAN_RADIO_TUNER_BAND_NONE  = 0;
 const uint8_t CAN_RADIO_TUNER_BAND_FM1   = 1;
